Now works with drupal 5
$query is now the first arg in xapian_do_search

Made undata path and dbpath optional

Added dbpath to settings page (drupal only)

Made xapian_do_search in to an object that has info and data.  

Info is the string returned by Query_get_description($query)
and data is an array with the filename, offset and length of the result.
The idea is to leave the formatting until later on - so xapian_do_search
will only deal with the search.   





git-svn-id: http://project.knowledgeforge.net/ukparse/svn/trunk@1556 9cc54934-f9f6-0310-8a8d-e8e977684441